requirements.txt
Flask
pytesseract
pillow
PyMuPDF
